predict-server/requirements-main.txt
#Django dependencies
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
idna==2.10
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
django-cors-headers==3.7.0

#DRF dependencies
django-rest-framework==0.1.0
djangorestframework==3.12.4

#Pandas dependencies
pandas==1.1.5
python-dateutil==2.8.1
six==1.15.0

#Postgresql
psycopg2==2.8.6
#Cassandra
cassandra-driver==3.25.0
#Spark
pyspark==3.1.1
#SQLAlchemy
sqlalchemy==1.4.7

#Django q
django-q==1.3.5
django-picklefield==3.0.1
ansicon==1.89.0
arrow==1.0.3
blessed==1.18.0
jinxed==1.1.0

#Postgresql
psycopg2==2.8.6

#Misc dependencies
requests==2.25.1
numpy==1.19.5
dill==0.3.3

#Cerberus vlidation
cerberus==1.3.4
